+++ Makefile
@@ -5,11 +5,12 @@ NAME = webserv
 GREEN=\033[32m
 RESET=\033[0m
 
-SOURCES = ./src/main.cpp ./src/server/Server.cpp  ./src/server/ServerUtils.cpp \
-					./src/server/RequestParser.cpp ./src/data_classes/MIME_type.cpp \
-					./src/data_classes/Statuscodes.cpp ./src/server/RequestUtils.cpp ./src/cfg/Config.cpp \
-					./src/server/ServerResponse.cpp ./src/server/Clients.cpp ./src/server/CGI.cpp \
-					./src/server/ServerManager.cpp ./src/server/ManagerUtils.cpp
+SOURCES = ./src/main.cpp ./src/server/Server.cpp  ./src/utils/ServerUtils.cpp \
+					./src/request/RequestParser.cpp ./src/data_classes/MIME_type.cpp \
+					./src/data_classes/Statuscodes.cpp ./src/request/RequestUtils.cpp ./src/cfg/Config.cpp \
+					./src/response/ErrorResponses.cpp ./src/server/Clients.cpp ./src/server/CGI.cpp \
+					./src/server/ServerManager.cpp ./src/utils/ManagerUtils.cpp ./src/response/ResponseMethods.cpp \
+					./src/response/ProcessResponse.cpp ./src/cfg/Locations.cpp ./src/utils/Utils.cpp
 
 OBJECTS = $(SOURCES:.cpp=.o)
