+++ config/main.conf
@@ -3,12 +3,12 @@ server
   	server_name exampleee.com;
 	root /Users/rnauke/Documents/GitHub/webserv/responseFiles;
   	index index.html index.htm index.nginx-debian.html;
-	listen 80;
+	listen 8080;
 	location / {
 		autoindex on;
 		index index.html;
-    max_client_body 10000;
-    deny_methods GET DELETE;
+    # max_client_body 1;
+    # deny_methods DELETE;
 		# random comment;
 		# haha
 		root /this/is/location/1;
@@ -23,6 +23,7 @@ server
 	location /Upload/ {
 		autoindex on;
 		index index.html;
+    # deny_methods DELETE;
 		# random comment;
 		# haha
 		root /this/is/location/1;
