+++ includes/RequestParser.hpp
@@ -6,7 +6,7 @@
 /*   By: njantsch <njantsch@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/14 17:22:44 by njantsch          #+#    #+#             */
-/*   Updated: 2024/02/13 18:59:07 by njantsch         ###   ########.fr       */
+/*   Updated: 2024/02/23 15:44:35 by njantsch         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -49,6 +49,7 @@ public:
   const std::string& getUri();
   const std::string& getHost();
   const std::string& getBody();
+  size_t             getBodySize();
   const std::string& getCurrdir();
   const std::string  getMapValue(const std::string key);
 };
