+++ includes/Utils.hpp
@@ -1,0 +1,22 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   Utils.hpp                                          :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: njantsch <njantsch@student.42.fr>          +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2024/02/23 15:48:15 by njantsch          #+#    #+#             */
+/*   Updated: 2024/02/23 15:51:53 by njantsch         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#pragma once
+
+#include <iostream>
+#include <sstream>
+
+short ft_stosh(const std::string& str);
+
+size_t ft_stosize(const std::string& str);
+
+std::string ft_itos(size_t num);
