+++ responseFiles/error413.html
@@ -1,0 +1,25 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Error 413</title>
+    <style>
+      body{
+      font-family: 'Segoe UI', Tahoma, Geneva, Verdana, sans-serif;
+      margin: 0;
+      padding: 0;
+      display: flex;
+      align-items: center;
+      justify-content: center;
+      height: 100vh;
+      }
+    </style>
+</head>
+<link rel="icon" type="image/x-icon" href="responseFiles/favicon.ico">
+<body>
+    <div class="wrapper">
+        <h1>Error 413 Payload Too Large!</h1>
+    </div>
+</body>
+</html>
