+++ src/cfg/Locations.cpp
@@ -1,56 +1,30 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ServerUtils.cpp                                    :+:      :+:    :+:   */
+/*   Locations.cpp                                      :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: njantsch <njantsch@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2024/01/17 14:35:15 by njantsch          #+#    #+#             */
-/*   Updated: 2024/02/23 12:49:45 by njantsch         ###   ########.fr       */
+/*   Created: 2024/02/23 15:37:26 by njantsch          #+#    #+#             */
+/*   Updated: 2024/02/23 16:29:34 by njantsch         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../includes/Server.hpp"
 
-int   Server::getFreeSocket()
+void  Server::initConfVars(Config& cfg)
 {
-  for (int i = 0; i < MAX_CLIENTS; i++)
-  {
-    if (this->_clientPollfds[i].fd == -1
-        && this->_clientDetails[i].getFdStatus() == UNUSED)
-      return (i);
-	}
-  return (-1);
-}
+  std::map<std::string, std::string> config = cfg.getConfig();
 
-void  Server::cleanUpClientFds()
-{
-  for (size_t i = 0; i < MAX_CLIENTS; i++) {
-    if (this->_clientDetails[i].getFdStatus() != UNUSED)
-      close(this->_clientPollfds[i].fd);
-  }
-}
-
-void  Server::removeFd(int i)
-{
-  if (close(this->_clientPollfds[i].fd) == -1)
-    perror("removeFd: close");
-  this->_clientDetails[i].cleanUp();
-  this->_clientDetails[i].cleanUpResponse();
-  this->_clientDetails[i].setFdStatus(UNUSED);
-  this->_clientDetails[i].setConStatus(KEEPALIVE);
-  this->_clientPollfds[i].fd = -1;
-  this->_clientPollfds[i].events = POLLIN;
-  this->_clientPollfds[i].revents = 0;
-  this->_nfds--;
-  std::cout << "Connection closed at idx: " << i << std::endl;
-}
-
-std::string Server::ft_itos(size_t num)
-{
-  std::ostringstream oss;
-  oss << num;
-  return (oss.str());
+  std::cout << "server port: " << ft_stosh(config.find("listen")->second) << std::endl;
+  this->_port = config.find("listen")->second;
+  this->_servername = config.find("server_name")->second;
+  this->_root = config.find("root")->second;
+  if (config.find("max_client_body") != config.end())
+    this->_maxClientBody = ft_stosize(config.find("max_client_body")->second);
+  else
+    this->_maxClientBody = LONG_MAX;
+  this->_locations = cfg.getLocations();
 }
 
 bool  Server::checkLocationPrelims(std::string method, size_t idx)
@@ -98,7 +72,3 @@ bool  Server::checkReferer(std::string method, size_t index)
   }
   return (false);
 }
-
-MIME_type& Server::getMimeType(void) {return (this->_data);}
-
-Statuscodes& Server::getStatuscodes(void) {return (this->_codes);}
