+++ src/utils/ManagerUtils.cpp
